+++ fltk-sys/src/input.rs
@@ -280,6 +280,12 @@ extern "C" {
         arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
     );
 }
+extern "C" {
+    pub fn Fl_Widget_visible(self_: *const Fl_Widget) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Widget_visible_r(self_: *const Fl_Widget) -> libc::c_int;
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Input {
@@ -539,6 +545,12 @@ extern "C" {
         arg1: *mut Fl_Input,
         arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
     );
+}
+extern "C" {
+    pub fn Fl_Input_visible(self_: *const Fl_Input) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Input_visible_r(self_: *const Fl_Input) -> libc::c_int;
 }
 extern "C" {
     pub fn Fl_Input_set_value(arg1: *mut Fl_Input, arg2: *const libc::c_char) -> libc::c_int;
@@ -898,6 +910,12 @@ extern "C" {
         arg1: *mut Fl_Int_Input,
         arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
     );
+}
+extern "C" {
+    pub fn Fl_Int_Input_visible(self_: *const Fl_Int_Input) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Int_Input_visible_r(self_: *const Fl_Int_Input) -> libc::c_int;
 }
 extern "C" {
     pub fn Fl_Int_Input_set_value(
@@ -1263,6 +1281,12 @@ extern "C" {
         arg1: *mut Fl_Float_Input,
         arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
     );
+}
+extern "C" {
+    pub fn Fl_Float_Input_visible(self_: *const Fl_Float_Input) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Float_Input_visible_r(self_: *const Fl_Float_Input) -> libc::c_int;
 }
 extern "C" {
     pub fn Fl_Float_Input_set_value(
@@ -1639,6 +1663,12 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Multiline_Input_visible(self_: *const Fl_Multiline_Input) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Multiline_Input_visible_r(self_: *const Fl_Multiline_Input) -> libc::c_int;
+}
+extern "C" {
     pub fn Fl_Multiline_Input_set_value(
         arg1: *mut Fl_Multiline_Input,
         arg2: *const libc::c_char,
@@ -2013,6 +2043,12 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Secret_Input_visible(self_: *const Fl_Secret_Input) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Secret_Input_visible_r(self_: *const Fl_Secret_Input) -> libc::c_int;
+}
+extern "C" {
     pub fn Fl_Secret_Input_set_value(
         arg1: *mut Fl_Secret_Input,
         arg2: *const libc::c_char,
@@ -2376,6 +2412,12 @@ extern "C" {
         arg1: *mut Fl_File_Input,
         arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
     );
+}
+extern "C" {
+    pub fn Fl_File_Input_visible(self_: *const Fl_File_Input) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_File_Input_visible_r(self_: *const Fl_File_Input) -> libc::c_int;
 }
 extern "C" {
     pub fn Fl_File_Input_set_down_box(self_: *mut Fl_File_Input, box_: libc::c_int);