+++ fltk-sys/src/table.rs
@@ -280,6 +280,12 @@ extern "C" {
         arg2: ::core::option::Option<unsafe extern "C" fn(arg1: *mut libc::c_void)>,
     );
 }
+extern "C" {
+    pub fn Fl_Widget_visible(self_: *const Fl_Widget) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Widget_visible_r(self_: *const Fl_Widget) -> libc::c_int;
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Group {
@@ -541,6 +547,18 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Group_visible(self_: *const Fl_Group) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Group_visible_r(self_: *const Fl_Group) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Group_current() -> *mut Fl_Group;
+}
+extern "C" {
+    pub fn Fl_Group_set_current(grp: *mut Fl_Group);
+}
+extern "C" {
     pub fn Fl_Group_begin(self_: *mut Fl_Group);
 }
 extern "C" {
@@ -557,6 +575,9 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Group_remove(self_: *mut Fl_Group, wid: *mut libc::c_void);
+}
+extern "C" {
+    pub fn Fl_Group_remove_by_index(self_: *mut Fl_Group, idx: libc::c_int);
 }
 extern "C" {
     pub fn Fl_Group_clear(self_: *mut Fl_Group);
@@ -831,6 +852,18 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Pack_visible(self_: *const Fl_Pack) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Pack_visible_r(self_: *const Fl_Pack) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Pack_spacing(self_: *mut Fl_Pack) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Pack_set_spacing(self_: *mut Fl_Pack, spacing: libc::c_int);
+}
+extern "C" {
     pub fn Fl_Pack_begin(self_: *mut Fl_Pack);
 }
 extern "C" {
@@ -847,6 +880,9 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Pack_remove(self_: *mut Fl_Pack, wid: *mut libc::c_void);
+}
+extern "C" {
+    pub fn Fl_Pack_remove_by_index(self_: *mut Fl_Pack, idx: libc::c_int);
 }
 extern "C" {
     pub fn Fl_Pack_clear(self_: *mut Fl_Pack);
@@ -1125,6 +1161,12 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Scroll_visible(self_: *const Fl_Scroll) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Scroll_visible_r(self_: *const Fl_Scroll) -> libc::c_int;
+}
+extern "C" {
     pub fn Fl_Scroll_scrollbar(self_: *mut Fl_Scroll) -> *mut libc::c_void;
 }
 extern "C" {
@@ -1162,6 +1204,9 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Scroll_remove(self_: *mut Fl_Scroll, wid: *mut libc::c_void);
+}
+extern "C" {
+    pub fn Fl_Scroll_remove_by_index(self_: *mut Fl_Scroll, idx: libc::c_int);
 }
 extern "C" {
     pub fn Fl_Scroll_clear(self_: *mut Fl_Scroll);
@@ -1436,6 +1481,12 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Tabs_visible(self_: *const Fl_Tabs) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Tabs_visible_r(self_: *const Fl_Tabs) -> libc::c_int;
+}
+extern "C" {
     pub fn Fl_Tabs_value(self_: *mut Fl_Tabs) -> *mut Fl_Widget;
 }
 extern "C" {
@@ -1479,6 +1530,9 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Tabs_remove(self_: *mut Fl_Tabs, wid: *mut libc::c_void);
+}
+extern "C" {
+    pub fn Fl_Tabs_remove_by_index(self_: *mut Fl_Tabs, idx: libc::c_int);
 }
 extern "C" {
     pub fn Fl_Tabs_clear(self_: *mut Fl_Tabs);
@@ -1753,6 +1807,12 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Tile_visible(self_: *const Fl_Tile) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Tile_visible_r(self_: *const Fl_Tile) -> libc::c_int;
+}
+extern "C" {
     pub fn Fl_Tile_begin(self_: *mut Fl_Tile);
 }
 extern "C" {
@@ -1769,6 +1829,9 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Tile_remove(self_: *mut Fl_Tile, wid: *mut libc::c_void);
+}
+extern "C" {
+    pub fn Fl_Tile_remove_by_index(self_: *mut Fl_Tile, idx: libc::c_int);
 }
 extern "C" {
     pub fn Fl_Tile_clear(self_: *mut Fl_Tile);
@@ -2047,6 +2110,12 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Wizard_visible(self_: *const Fl_Wizard) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Wizard_visible_r(self_: *const Fl_Wizard) -> libc::c_int;
+}
+extern "C" {
     pub fn Fl_Wizard_next(arg1: *mut Fl_Wizard);
 }
 extern "C" {
@@ -2075,6 +2144,9 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Wizard_remove(self_: *mut Fl_Wizard, wid: *mut libc::c_void);
+}
+extern "C" {
+    pub fn Fl_Wizard_remove_by_index(self_: *mut Fl_Wizard, idx: libc::c_int);
 }
 extern "C" {
     pub fn Fl_Wizard_clear(self_: *mut Fl_Wizard);
@@ -2364,6 +2436,12 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Color_Chooser_visible(self_: *const Fl_Color_Chooser) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Color_Chooser_visible_r(self_: *const Fl_Color_Chooser) -> libc::c_int;
+}
+extern "C" {
     pub fn Fl_Color_Chooser_r(self_: *mut Fl_Color_Chooser) -> f64;
 }
 extern "C" {
@@ -2398,6 +2476,9 @@ extern "C" {
     pub fn Fl_Color_Chooser_remove(self_: *mut Fl_Color_Chooser, wid: *mut libc::c_void);
 }
 extern "C" {
+    pub fn Fl_Color_Chooser_remove_by_index(self_: *mut Fl_Color_Chooser, idx: libc::c_int);
+}
+extern "C" {
     pub fn Fl_Color_Chooser_clear(self_: *mut Fl_Color_Chooser);
 }
 extern "C" {
@@ -2411,12 +2492,6 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Color_Chooser_resizable(self_: *mut Fl_Color_Chooser, arg1: *mut libc::c_void);
-}
-extern "C" {
-    pub fn Fl_Pack_spacing(self_: *mut Fl_Pack) -> libc::c_int;
-}
-extern "C" {
-    pub fn Fl_Pack_set_spacing(self_: *mut Fl_Pack, spacing: libc::c_int);
 }
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
@@ -2679,6 +2754,12 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Table_visible(self_: *const Fl_Table) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Table_visible_r(self_: *const Fl_Table) -> libc::c_int;
+}
+extern "C" {
     pub fn Fl_Table_begin(self_: *mut Fl_Table);
 }
 extern "C" {
@@ -2695,6 +2776,9 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Table_remove(self_: *mut Fl_Table, wid: *mut libc::c_void);
+}
+extern "C" {
+    pub fn Fl_Table_remove_by_index(self_: *mut Fl_Table, idx: libc::c_int);
 }
 extern "C" {
     pub fn Fl_Table_clear(self_: *mut Fl_Table);
@@ -3214,6 +3298,12 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Table_Row_visible(self_: *const Fl_Table_Row) -> libc::c_int;
+}
+extern "C" {
+    pub fn Fl_Table_Row_visible_r(self_: *const Fl_Table_Row) -> libc::c_int;
+}
+extern "C" {
     pub fn Fl_Table_Row_begin(self_: *mut Fl_Table_Row);
 }
 extern "C" {
@@ -3230,6 +3320,9 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Table_Row_remove(self_: *mut Fl_Table_Row, wid: *mut libc::c_void);
+}
+extern "C" {
+    pub fn Fl_Table_Row_remove_by_index(self_: *mut Fl_Table_Row, idx: libc::c_int);
 }
 extern "C" {
     pub fn Fl_Table_Row_clear(self_: *mut Fl_Table_Row);