+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+
+## [0.15.6] - 2021-03-03
+### Changes
+- Add screen scale getters and setters in the app module.
+- add WidgetExt::visible() and visible_r() (recursive) methods.
+- Add GroupExt::remove_by_index().
+- MenuItem::popup() no longer requires mut self.
+- Update deps.
+
 ## [0.15.5] - 2021-03-01
 ### Changes
 - Check for 0 children with Pack::auto_layout(). Thanks @dheijl.